Adds min height to new chat
