+++ vector/src/main/java/im/vector/app/features/spaces/SpaceListBottomSheet.kt
@@ -27,6 +27,7 @@ import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import im.vector.app.R
 import im.vector.app.core.extensions.replaceChildFragment
+import im.vector.app.core.extensions.setPeekHeightAsScreenPercentage
 import im.vector.app.databinding.FragmentSpacesBottomSheetBinding
 
 class SpaceListBottomSheet : BottomSheetDialogFragment() {
@@ -46,14 +47,6 @@ class SpaceListBottomSheet : BottomSheetDialogFragment() {
         return super.onCreateDialog(savedInstanceState).apply {
             (this as BottomSheetDialog).setPeekHeightAsScreenPercentage(0.75f)
         }
-    }
-
-    @Suppress("DEPRECATION")
-    private fun BottomSheetDialog.setPeekHeightAsScreenPercentage(@FloatRange(from = 0.0, to = 1.0) percentage: Float) {
-        val displayMetrics = DisplayMetrics()
-        window?.windowManager?.defaultDisplay?.getMetrics(displayMetrics)
-        val height = displayMetrics.heightPixels
-        behavior.setPeekHeight((height * percentage).toInt(), true)
     }
 
     companion object {
